+++ src/Keybinds/KeybindManager.cpp
@@ -93,22 +93,3 @@ bool KeybindManager::processMSG(KeyState state)
 #include <Geode/loader/SettingV3.hpp>
 
 using namespace geode::prelude;
-
-$on_mod(Loaded)
-{
-    listenForKeybindSettingPresses("open-menu-keybind", [](Keybind const& keybind, bool down, bool repeat, double timestamp)
-    {
-        if (down && !repeat)
-        {
-            if (AndroidUI::get())
-            {
-                // close one menu so that option displays are gone before the main popup
-                CCKeyboardDispatcher::get()->dispatchKeyboardMSG(enumKeyCodes::KEY_Escape, true, false, 0);
-            }            
-            else
-            {
-                AndroidUI::addToScene();
-            }
-        }
-});
-};