+++ src/Labels/Nodes/ImageNode.cpp
@@ -18,6 +18,7 @@ void ImageNode::labelConfigUpdated()
     if (!sprite)
         return;
 
+    sprite->setOpacity(config.opacity);
     sprite->setAnchorPoint(ccp(0, 0));
     this->addChild(sprite);
 }